+++ NEWS.md
@@ -14,9 +14,13 @@ dbFetch(res, 5000)
 ```
 
 * When creating/appending partitions to a table, `dbWriteTable` opts to use `alter table` instead of standard `msck repair table`. This is to improve performance when appending to tables with high number of existing partitions.
+* `dbWriteTable` now allows json to be appended to json ddls created with the Openx-JsonSerDe library.
 
 ## Bug
 * `dbWriteTable` would throw `throttling error` every now and again, `retry_api_call` as been built to handle the parsing of data between R and AWS S3.
+
+## Documentation
+* `dbWriteTable` added support ddl structures for user who have created ddl's outside of `noctua`
 
 # noctua 1.6.0
 ## New Feature